refactor mypage & create admin, fix menu
