+++ frontend/src/App.jsx
@@ -44,7 +44,7 @@ const App = () => {
           <Route path="/login" element={<LoginForm />} />
           <Route path="/register" element={<RegisterForm />} />
           <Route path="/mypage" element={<MyPage />} />
-          <Route path="/client" element={<ClientForm />} />
+          <Route path="/addclient" element={<ClientForm />} />
           <Route path="/client/:id" element={<Client />} />
           <Route path="/client/:id/update" element={<UpdateClient />} />
           <Route path="/client/:id/salaryform" element={<SalaryForm />} />
